Ensure domain names & SANs are lower-case
If you provide an upper-case domain name, this module fails with the following error:

  on .terraform/modules/my_certificate/main.tf line 37, in resource "aws_route53_record" "default":
  37:   name            = each.value.name
A reference to "each.value" has been used in a context in which it
unavailable, such as when the configuration no longer contains the value in
its "for_each" expression. Remove this reference to each.value in your
configuration to work around this error.